feat!: support esm usage of eslint flat config and drop commonjs
