+++ eslint.config.js
@@ -1,0 +1,3 @@
+import config from './eslint/index.js';
+
+export default config(['node']);
