--- eslint.config.mjs
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/extensions -- fix it after refactor to esm
-import config from './eslint/flat/config.js';
-
-export default config('node');
