+++ eslint/browser.js
@@ -1,6 +1,13 @@
-module.exports = {
-    extends: require.resolve('./internals/common'),
-    env: {
-        browser: true,
+import globals from 'globals';
+
+/**
+ * @returns {import('typescript-eslint').ConfigArray}
+ */
+export default [
+    {
+        files: ['**/*.{js,mjs,mts,cjs,cts}'],
+        languageOptions: {
+            globals: globals.browser,
+        },
     },
-};
+];
