--- eslint/flat/presets/browser.js
@@ -1,13 +1,0 @@
-const globals = require('globals');
-
-/**
- * @returns {import('typescript-eslint').ConfigArray}
- */
-module.exports = [
-    {
-        files: ['**/*.{js,mjs,mts,cjs,cts}'],
-        languageOptions: {
-            globals: globals.browser,
-        },
-    },
-];
