--- eslint/flat/presets/react.js
@@ -1,33 +1,0 @@
-const globals = require('globals');
-const reactPlugin = require('eslint-plugin-react');
-const reactHooks = require('eslint-plugin-react-hooks');
-const importPlugin = require('eslint-plugin-import');
-const reactRules = require('../../react').rules;
-
-/**
- * @returns {import('typescript-eslint').ConfigArray}
- */
-module.exports = [
-    {
-        files: ['**/*.{js,mjs,cjs,jsx,mjsx,ts,tsx,mtsx}'],
-        settings: {
-            ...importPlugin.flatConfigs.react.settings,
-            react: {version: 'detect'},
-        },
-        languageOptions: {
-            ...reactPlugin.configs.flat.recommended.languageOptions,
-            ...importPlugin.flatConfigs.react.languageOptions,
-            globals: {
-                ...globals.browser,
-            },
-        },
-        plugins: {
-            react: reactPlugin,
-            'react-hooks': reactHooks,
-        },
-        rules: {
-            ...reactHooks.configs.recommended.rules,
-            ...reactRules,
-        },
-    },
-];
