+++ eslint/index.js
@@ -1,7 +1,31 @@
-const build = require('./flat/build');
-const config = require('./flat/config');
+import base from './base.js';
+import browser from './browser.js';
+import node from './node.js';
+import react from './react.js';
+import typescript from './typescript.js';
 
-module.exports.flatConfigs = {
-    build,
-    config,
+/**
+ * @typedef {'node' | 'browser' | 'react' | 'typescript'} Preset
+ */
+
+/**
+ * Create eslint flat config
+ *
+ * @param {Array<Preset> | Preset | undefined} presets
+ * @param {import('typescript-eslint').ConfigArray} extendedFlatConfigs
+ * @returns {import('typescript-eslint').ConfigArray}
+ */
+export default (presets, ...extendedFlatConfigs) => {
+    /**
+     * @type {Preset[]}
+     */
+    const resolvedPresets = Array.isArray(presets ?? []) ? presets : [presets];
+    return [
+        ...base,
+        ...resolvedPresets.includes('node') ? node : [],
+        ...resolvedPresets.includes('browser') ? browser : [],
+        ...resolvedPresets.includes('react') ? react : [],
+        ...resolvedPresets.includes('typescript') ? typescript : [],
+        ...extendedFlatConfigs,
+    ];
 };
