--- eslint/internals/base.js
@@ -1,261 +1,0 @@
-module.exports = {
-    rules: {
-        'no-redeclare': ['off'],
-        'max-lines': [
-            'warn',
-            {
-                max: 300,
-                skipBlankLines: true,
-                skipComments: true,
-            },
-        ],
-        'max-lines-per-function': [
-            'warn',
-            {
-                max: 100,
-                skipBlankLines: true,
-                skipComments: true,
-            },
-        ],
-        'no-duplicate-imports': ['off'],
-        'no-empty-function': [
-            'off',
-            {
-                allow: ['generatorFunctions', 'methods', 'generatorMethods', 'setters', 'asyncMethods'],
-            },
-        ],
-        'no-unused-vars': ['off'],
-        'init-declarations': ['off'],
-        'no-undef': ['off'],
-        'require-await': ['off'],
-        'no-use-before-define': [
-            'off',
-            {
-                variables: false,
-                classes: false,
-                functions: true,
-            },
-        ],
-        'no-dupe-class-members': ['off'],
-        'no-useless-constructor': ['off'],
-        'dot-notation': [
-            'off',
-            {
-                allowKeywords: true,
-                allowPattern: '',
-            },
-        ],
-        'no-void': [
-            'error',
-            {
-                allowAsStatement: true,
-            },
-        ],
-        'accessor-pairs': ['error'],
-        'array-callback-return': ['error'],
-        'arrow-body-style': ['off'],
-        'block-scoped-var': ['warn'],
-        'callback-return': ['off'],
-        camelcase: ['error'],
-        'capitalized-comments': ['off'],
-        'class-methods-use-this': ['off'],
-        complexity: ['warn', 10],
-        'consistent-return': ['off'],
-        'consistent-this': ['off'],
-        'constructor-super': ['error'],
-        curly: ['error', 'all'],
-        'default-case': ['off'],
-        eqeqeq: [
-            'error',
-            'always',
-            {
-                null: 'ignore',
-            },
-        ],
-        'for-direction': ['off'],
-        'func-name-matching': ['off'],
-        'func-names': ['warn', 'always'],
-        'func-style': ['off'],
-        'global-require': ['warn'],
-        'guard-for-in': ['warn'],
-        'handle-callback-err': ['off'],
-        'id-blacklist': ['off'],
-        'id-length': ['off'],
-        'id-match': ['off'],
-        'indent-legacy': ['off'],
-        'max-depth': ['warn', 3],
-        'max-nested-callbacks': ['off'],
-        'max-params': ['off'],
-        'max-statements': [
-            'warn',
-            30,
-            {
-                ignoreTopLevelFunctions: true,
-            },
-        ],
-        'new-cap': ['error'],
-        'no-alert': ['warn'],
-        'no-array-constructor': ['off'],
-        'no-await-in-loop': ['off'],
-        'no-bitwise': ['off'],
-        'no-buffer-constructor': ['warn'],
-        'no-caller': ['error'],
-        'no-case-declarations': ['error'],
-        'no-catch-shadow': ['off'],
-        'no-class-assign': ['error'],
-        'no-compare-neg-zero': ['error'],
-        'no-cond-assign': ['error'],
-        'no-console': [
-            'error',
-            {
-                allow: ['error', 'warn', 'info'],
-            },
-        ],
-        'no-const-assign': ['error'],
-        'no-constant-condition': ['error'],
-        'no-continue': ['off'],
-        'no-control-regex': ['error'],
-        'no-debugger': ['error'],
-        'no-delete-var': ['error'],
-        'no-div-regex': ['off'],
-        'no-dupe-args': ['error'],
-        'no-dupe-keys': ['error'],
-        'no-duplicate-case': ['error'],
-        'no-else-return': ['off'],
-        'no-empty': ['error'],
-        'no-empty-character-class': ['error'],
-        'no-empty-pattern': ['error'],
-        'no-eq-null': ['off'],
-        'no-eval': ['error'],
-        'no-ex-assign': ['error'],
-        'no-extend-native': ['error'],
-        'no-extra-bind': ['warn'],
-        'no-extra-boolean-cast': ['error'],
-        'no-extra-label': ['error'],
-        'no-fallthrough': ['error'],
-        'no-func-assign': ['warn'],
-        'no-global-assign': ['error'],
-        'no-implicit-coercion': ['off'],
-        'no-implicit-globals': ['off'],
-        'no-implied-eval': ['error'],
-        'no-inline-comments': ['off'],
-        'no-inner-declarations': ['error'],
-        'no-invalid-regexp': ['error'],
-        'no-invalid-this': ['off'],
-        'no-irregular-whitespace': ['error'],
-        'no-iterator': ['off'],
-        'no-label-var': ['off'],
-        'no-labels': ['off'],
-        'no-lone-blocks': ['off'],
-        'no-lonely-if': ['error'],
-        'no-loop-func': ['warn'],
-        'no-magic-numbers': ['off'],
-        'no-mixed-requires': ['off'],
-        'no-multi-assign': ['warn'],
-        'no-multi-str': ['warn'],
-        'no-negated-condition': ['warn'],
-        'no-negated-in-lhs': ['error'],
-        'no-nested-ternary': ['off'],
-        'no-new': ['error'],
-        'no-new-func': ['error'],
-        'no-new-object': ['error'],
-        'no-new-require': ['error'],
-        'no-new-symbol': ['error'],
-        'no-new-wrappers': ['error'],
-        'no-obj-calls': ['error'],
-        'no-octal': ['error'],
-        'no-octal-escape': ['error'],
-        'no-param-reassign': [
-            'warn',
-            {
-                props: true,
-            },
-        ],
-        'no-path-concat': ['off'],
-        'no-plusplus': ['off'],
-        'no-process-env': ['off'],
-        'no-process-exit': ['off'],
-        'no-proto': ['error'],
-        'no-prototype-builtins': ['off'],
-        'no-regex-spaces': ['error'],
-        'no-restricted-globals': ['off'],
-        'no-restricted-modules': ['off'],
-        'no-restricted-properties': ['off'],
-        'no-restricted-syntax': ['off'],
-        'no-return-assign': ['error', 'except-parens'],
-        'no-return-await': ['error'],
-        'no-script-url': ['warn'],
-        'no-self-assign': ['error'],
-        'no-self-compare': ['error'],
-        'no-sequences': ['error'],
-        'no-shadow': ['off'],
-        'no-shadow-restricted-names': ['off'],
-        'no-sparse-arrays': ['error'],
-        'no-sync': ['off'],
-        'no-template-curly-in-string': ['off'],
-        'no-ternary': ['off'],
-        'no-this-before-super': ['error'],
-        'no-throw-literal': ['warn'],
-        'no-undef-init': ['warn'],
-        'no-undefined': ['off'],
-        'no-underscore-dangle': ['warn'],
-        'no-unexpected-multiline': ['error'],
-        'no-unmodified-loop-condition': ['warn'],
-        'no-unneeded-ternary': ['error'],
-        'no-unreachable': ['error'],
-        'no-unreachable-loop': ['error'],
-        'no-unsafe-finally': ['error'],
-        'no-unsafe-negation': ['error'],
-        'no-unused-expressions': [
-            'warn',
-            {
-                allowShortCircuit: true,
-                allowTernary: false,
-                allowTaggedTemplates: false,
-                enforceForJSX: false,
-            },
-        ],
-        'no-unused-labels': ['error'],
-        'no-useless-call': ['warn'],
-        'no-useless-computed-key': ['error'],
-        'no-useless-concat': ['warn'],
-        'no-useless-escape': ['error'],
-        'no-useless-rename': ['error'],
-        'no-useless-return': ['error'],
-        'no-var': ['error'],
-        'no-warning-comments': ['off'],
-        'no-with': ['error'],
-        'object-shorthand': ['off', 'consistent-as-needed'],
-        'one-var': ['error', 'never'],
-        'operator-assignment': ['off'],
-        'prefer-arrow-callback': [
-            'warn',
-            {
-                allowNamedFunctions: true,
-                allowUnboundThis: true,
-            },
-        ],
-        'prefer-const': ['error'],
-        'prefer-destructuring': ['off'],
-        'prefer-numeric-literals': ['error'],
-        'prefer-promise-reject-errors': ['warn'],
-        'prefer-reflect': ['off'],
-        'prefer-rest-params': ['error'],
-        'prefer-spread': ['warn'],
-        'prefer-template': ['off'],
-        radix: ['error'],
-        'require-jsdoc': ['off'],
-        'require-yield': ['error'],
-        'sort-imports': ['off'],
-        'sort-keys': ['off'],
-        'sort-vars': ['off'],
-        strict: ['off'],
-        'symbol-description': ['warn'],
-        'unicode-bom': ['warn'],
-        'use-isnan': ['error'],
-        'valid-jsdoc': ['off'],
-        'valid-typeof': ['error'],
-        'vars-on-top': ['off'],
-        yoda: ['warn'],
-    },
-};
