+++ lint-staged.config.js
